+++ internal/logger/logger_linux.go
@@ -4,19 +4,11 @@ package logger
 
 import (
 	"os"
-	"unsafe"
 
 	"golang.org/x/sys/unix"
 )
 
 const SupportsColorEscapes = true
-
-type winsize struct {
-	ws_row    uint16
-	ws_col    uint16
-	ws_xpixel uint16
-	ws_ypixel uint16
-}
 
 func GetTerminalInfo(file *os.File) (info TerminalInfo) {
 	fd := file.Fd()
@@ -27,9 +19,8 @@ func GetTerminalInfo(file *os.File) (info TerminalInfo) {
 		info.UseColorEscapes = true
 
 		// Get the width of the window
-		w := new(winsize)
-		if _, _, err := unix.Syscall(unix.SYS_IOCTL, fd, unix.TIOCGWINSZ, uintptr(unsafe.Pointer(w))); err == 0 {
-			info.Width = int(w.ws_col)
+		if w, err := unix.IoctlGetWinsize(int(fd), unix.TIOCGWINSZ); err == nil {
+			info.Width = int(w.Col)
 		}
 	}